데이터수집_공유/requirements.txt
opencv-python
opencv-contrib-python
numpy
mediapipe==0.10.14
Pillow
